requirements.txt
nltk
spacy
flask
python-dateutil
email-validator
googletrans-py
json5
requests
python-dotenv
pytest
pytest-mock
pypdf
python-docx
sentence-transformers
faiss-cpu
PyMuPDF
